Add files via upload
This is one of my frontend practice projects. I tried to clone a website layout to build mine. The only problem is that it's not responsive and most projects' websites I built were never responsive.Please does anyone out there know the best ways or practices to make websites responsive?